ph_image_thread: same indent
I'm trying not to conjure a tabs vs. spaces debate here.
The indents were clearly 8 in width for the spaces lines, however the rest of the file is mostly 4 in width where indent is spaces.
Using hard tabs instead of soft tabs simply side-steps the entire "4 or 8" debate and also is an accessibility feature.

Signed-off-by: benaryorg <binary@benary.org>